# docs/specs/data-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Outage Explorer Data API",
    "version": "1.0.0",
    "description": "Implemented HTTP contract for explicitly configured services. Synthetic examples; Linux runtime validation belongs to the user and live resources are not enabled."
  },
  "paths": {
...
    "/api/refresh": {
      "post": {
        "operationId": "admitRefresh",
        "summary": "admitRefresh",
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RefreshReceipt",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshReceipt"
                }
              }
            }
          },
          "400": {
            "description": "Error400",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400"
                }
              }
            }
          },
          "401": {
            "description": "Error401",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error401"
                }
              }
            }
          },
          "403": {
            "description": "Error403",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "409": {
            "description": "Error409",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error409"
                }
              }
            }
          },
          "503": {
            "description": "Error503",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error503"
                }
              }
            }
          },
          "202": {
            "description": "RefreshReceipt",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "schema": {
                  "const": "3"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshReceipt"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Requires current session and application authorization. All responses use no-store. Duplicate/unknown parameters and unsupported bodies are rejected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        }
      }
    },
    "/api/refresh/latest": {
      "get": {
        "operationId": "latestRefresh",
        "summary": "latestRefresh",
        "x-implementation-status": "implemented",
        "parameters": [],
        "responses": {
          "200": {
            "description": "LatestRefresh",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestRefresh"
                }
              }
            }
          },
          "401": {
            "description": "Error401",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error401"
                }
              }
            }
          },
          "403": {
            "description": "Error403",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "503": {
            "description": "Error503",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error503"
                }
              }
            }
          },
          "400": {
            "description": "Error400",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Requires current session and application authorization. All responses use no-store. Duplicate/unknown parameters and unsupported bodies are rejected."
      }
    },
    "/api/refresh/{run_id}": {
      "get": {
        "operationId": "getRefresh",
        "summary": "getRefresh",
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RefreshRun",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshRun"
                }
              }
            }
          },
          "400": {
            "description": "Error400",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error400"
                }
              }
            }
          },
          "401": {
            "description": "Error401",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error401"
                }
              }
            }
          },
          "403": {
            "description": "Error403",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error403"
                }
              }
            }
          },
          "404": {
            "description": "Error404",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error404"
                }
              }
            }
          },
          "503": {
            "description": "Error503",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                },
                "description": "Bounded retry suggestion, agrees with error.retry_after_seconds when present."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error503"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Requires current session and application authorization. All responses use no-store. Duplicate/unknown parameters and unsupported bodies are rejected."
      }
    }
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-outage_session",
        "description": "Production HTTPS cookie. Explicit local HTTP mode uses outage_session; reuse the configured auth transport."
      }
    },
    "schemas": {
...
      "Coverage": {
        "type": "object",
        "properties": {
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ]
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "start_date",
          "end_date"
        ],
        "additionalProperties": false
      },
      "Interval": {
        "type": "object",
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date"
          },
          "end_date": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "start_date",
          "end_date"
        ],
        "additionalProperties": false
      },
...
      "RefreshRequest": {
        "type": "object",
        "properties": {},
        "required": [],
        "additionalProperties": false
      },
      "RefreshReceipt": {
        "type": "object",
        "properties": {
          "run_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "running",
              "succeeded",
              "retained",
              "failed",
              "interrupted",
              "publication_unknown"
            ]
          },
          "effective_interval": {
            "$ref": "#/components/schemas/Interval"
          },
          "status_url": {
            "type": "string",
            "pattern": "^/api/refresh/[^/]+$"
          }
        },
        "required": [
          "run_id",
          "status",
          "effective_interval",
          "status_url"
        ],
        "additionalProperties": false
      },
      "Quality": {
        "type": "object",
        "properties": {
          "received_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "selected_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "excluded_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "duplicates_collapsed": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "superseded_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "retained_invalid_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "retained_absent_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "carried_outside_interval_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "modeled_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "retained_entire_dataset": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ]
          },
          "exclusion_reasons": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "count": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "required": [
                "code",
                "count"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "received_rows",
          "selected_rows",
          "excluded_rows",
          "duplicates_collapsed",
          "superseded_rows",
          "retained_invalid_rows",
          "retained_absent_rows",
          "carried_outside_interval_rows",
          "modeled_rows",
          "retained_entire_dataset",
          "exclusion_reasons"
        ],
        "additionalProperties": false
      },
      "RefreshDataset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "national",
              "facilities",
              "generators"
            ]
          },
          "status": {
            "enum": [
              "pending",
              "processing",
              "succeeded",
              "retained",
              "failed"
            ]
          },
          "coverage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coverage"
              },
              {
                "type": "null"
              }
            ]
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Quality"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "coverage",
          "quality"
        ],
        "additionalProperties": false
      },
      "SafeFailure": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "Publication": {
        "type": "object",
        "properties": {
          "state": {
            "enum": [
              "pending",
              "published",
              "not_published",
              "unknown"
            ]
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ]
          },
          "previous_generation_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ]
          },
          "no_publication_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "state",
          "generation_id",
          "previous_generation_id",
          "no_publication_reason"
        ],
        "additionalProperties": false
      },
      "RefreshRun": {
        "type": "object",
        "properties": {
          "run_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "running",
              "succeeded",
              "retained",
              "failed",
              "interrupted",
              "publication_unknown"
            ]
          },
          "stage": {
            "enum": [
              "queued",
              "retrieving",
              "modeling",
              "persisting",
              "verifying",
              "publishing",
              "finished"
            ]
          },
          "effective_interval": {
            "$ref": "#/components/schemas/Interval"
          },
          "accepted_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "Z$"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "Z$"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "Z$"
              },
              {
                "type": "null"
              }
            ]
          },
          "datasets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefreshDataset"
            },
            "minItems": 3,
            "maxItems": 3
          },
          "publication": {
            "$ref": "#/components/schemas/Publication"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SafeFailure"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "status",
          "stage",
          "effective_interval",
          "accepted_at",
          "started_at",
          "finished_at",
          "datasets",
          "publication",
          "failure"
        ],
        "additionalProperties": false
      },
      "LatestRefresh": {
        "type": "object",
        "properties": {
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RefreshRun"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run"
        ],
        "additionalProperties": false
      },
      "Error400": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "invalid_request",
                  "invalid_sql",
                  "unsupported_sql",
                  "page_out_of_range",
                  "page_size_mismatch"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error401": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "unauthenticated"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error403": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "forbidden"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error404": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "dataset_unavailable",
                  "query_unavailable",
                  "refresh_unavailable"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Error409": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "refresh_busy",
                  "idempotency_conflict"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Error503": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "enum": [
                  "query_busy",
                  "result_capacity_exhausted",
                  "data_unavailable",
                  "service_unavailable"
                ]
              },
              "message": {
                "type": "string"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1,
                "maximum": 3600
              },
              "details": {
                "type": "object",
                "properties": {
                  "query_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "Z$"
                  }
                },
                "required": [
                  "query_id",
                  "expires_at"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}